Figured out why the price is being floored (I have dum) but now it's starting with two of the things.

+++ src/Utillity.js
@@ -5,7 +5,7 @@ export function format(num) {
 }
 
 export function numPriceFromFormattedPrice(price) {
-	return parseInt(price.replace("£", ""));
+	return parseFloat(price.replace("£", ""));
 }
 
 export function formatPOIDFromNum(poIDNum) {
